backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wireframe"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "pydantic>=2.9",
    "pydantic-settings>=2.0",
    "langgraph>=0.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "structlog",
    "httpx",
    "marker-pdf",
    "kicad-sch-api",
    "vcrpy",
    "openai>=1.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "arq>=0.25",
    "redis>=5.0",
    "instructor>=1.0",
    "langgraph-checkpoint-postgres>=1.0",
    "psycopg[binary]>=3.1",
    "dacite>=1.8",
    "docling",
    "psycopg2-binary",
]

[tool.hatch.build.targets.wheel]
packages = ["wireframe"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "N", "S", "PLE", "PLW", "RUF", "TID"]
ignore = []
